// src/modules/product/redux/product.types.ts
export enum ProductTypes {
  PRODUCT_ACTION_START = "PRODUCT_ACTION_START",
  PRODUCT_GET_SUCCESS = "PRODUCT_GET_SUCCESS",
  PRODUCT_GET_FAILED = "PRODUCT_GET_FAILED",
  PRODUCT_CREATE_SUCCESS = "PRODUCT_CREATE_SUCCESS",
  PRODUCT_CREATE_FAILED = "PRODUCT_CREATE_FAILED",
  PRODUCT_GET_BY_ID_SUCCESS = "PRODUCT_GET_BY_ID_SUCCESS",
  PRODUCT_GET_BY_ID_FAILED = "PRODUCT_GET_BY_ID_FAILED",
  PRODUCT_UPDATE_SUCCESS = "PRODUCT_UPDATE_SUCCESS",
  PRODUCT_DELETE_SUCCESS = "PRODUCT_DELETE_SUCCESS",
  PRODUCT_DELETE_FAILED = "PRODUCT_DELETE_FAILED",
  PRODUCT_UPDATE_FAILED = "PRODUCT_UPDATE_FAILED",
  PRODUCT_ACTION_END = "PRODUCT_ACTION_END",
}